Fix panic caused by race condition when writing while dropping reader
`Writer` has a race condition that might cause a panic if `poll_write` is called more than once while the `Reader` is in the processed of being dropped on another thread.

The race condition is caused by the drop order of `Reader`: Since fields are dropped in the order they are defined in, the buffer pool channel is dropped _before_ the buffer stream channel. If `poll_write` is called more than once between these two drop calls, then the underlying channel will panic because we've polled it after already returning `None`.

The fix is to ensure the buffer stream channel is closed _first_ before the buffer pool channel, since that is the channel we use to detect when the reader is closed. One way of doing this would be to just reverse the order these two fields are defined in, but it is better to be clear that the drop order matters by using an explicit drop handler.

See also https://github.com/sagebind/isahc/issues/295.

+++ tests/pipe.rs
@@ -90,7 +90,7 @@ fn pipe_lots_of_data() {
 }
 
 #[quickcheck]
-fn read_write_chunks_random(chunks: u16) {
+fn read_write_chunks_random(chunks: u8) {
     block_on(async {
         let data = [0; 8192];
         let (mut reader, mut writer) = pipe();
